cmd/ddltest: migrate test-infra to testify
