+++ cmd/ddltest/column_serial_test.go
@@ -19,6 +19,7 @@ import (
 	"reflect"
 	"sync"
 	"sync/atomic"
+	"testing"
 	"time"
 
 	. "github.com/pingcap/check"
@@ -30,17 +31,18 @@ import (
 	"github.com/pingcap/tidb/table"
 	"github.com/pingcap/tidb/table/tables"
 	"github.com/pingcap/tidb/types"
+	"github.com/stretchr/testify/require"
 	"go.uber.org/zap"
 	goctx "golang.org/x/net/context"
 )
 
 // After add column finished, check the records in the table.
-func (s *TestDDLSuite) checkAddColumn(c *C, rowID int64, defaultVal interface{}, updatedVal interface{}) {
+func (s *ddlSuite) checkAddColumn(t *testing.T, rowID int64, defaultVal interface{}, updatedVal interface{}) {
 	ctx := s.ctx
 	err := ctx.NewTxn(goctx.Background())
-	c.Assert(err, IsNil)
+	require.NoError(t, err)
 
-	tbl := s.getTable(c, "test_column")
+	tbl := s.getTable(t, "test_column")
 	oldInsertCount := int64(0)
 	newInsertCount := int64(0)
 	oldUpdateCount := int64(0)
@@ -75,24 +77,24 @@ func (s *TestDDLSuite) checkAddColumn(c *C, rowID int64, defaultVal interface{},
 
 		return true, nil
 	})
-	c.Assert(err, IsNil)
+	require.NoError(t, err)
 
 	deleteCount := rowID - oldInsertCount - newInsertCount - oldUpdateCount - newUpdateCount
-	c.Assert(oldInsertCount, GreaterEqual, int64(0))
-	c.Assert(newInsertCount, GreaterEqual, int64(0))
-	c.Assert(oldUpdateCount, Greater, int64(0))
-	c.Assert(newUpdateCount, Greater, int64(0))
-	c.Assert(deleteCount, Greater, int64(0))
+	require.GreaterOrEqual(t, oldInsertCount, int64(0))
+	require.GreaterOrEqual(t, newInsertCount, int64(0))
+	require.Greater(t, oldUpdateCount, int64(0))
+	require.Greater(t, newUpdateCount, int64(0))
+	require.Greater(t, deleteCount, int64(0))
 }
 
-func (s *TestDDLSuite) checkDropColumn(c *C, rowID int64, alterColumn *table.Column, updateDefault interface{}) {
+func (s *ddlSuite) checkDropColumn(t *testing.T, rowID int64, alterColumn *table.Column, updateDefault interface{}) {
 	ctx := s.ctx
 	err := ctx.NewTxn(goctx.Background())
-	c.Assert(err, IsNil)
+	require.NoError(t, err)
 
-	tbl := s.getTable(c, "test_column")
+	tbl := s.getTable(t, "test_column")
 	for _, col := range tbl.Cols() {
-		c.Assert(col.ID, Not(Equals), alterColumn.ID)
+		require.NotEqual(t, alterColumn.ID, col.ID)
 	}
 	insertCount := int64(0)
 	updateCount := int64(0)
@@ -108,15 +110,18 @@ func (s *TestDDLSuite) checkDropColumn(c *C, rowID int64, alterColumn *table.Col
 		}
 		return true, nil
 	})
-	c.Assert(err, IsNil)
+	require.NoError(t, err)
 
 	deleteCount := rowID - insertCount - updateCount
-	c.Assert(insertCount, Greater, int64(0))
-	c.Assert(updateCount, Greater, int64(0))
-	c.Assert(deleteCount, Greater, int64(0))
+	require.Greater(t, insertCount, int64(0))
+	require.Greater(t, updateCount, int64(0))
+	require.Greater(t, deleteCount, int64(0))
 }
 
-func (s *TestDDLSuite) TestColumn(c *C) {
+func TestColumn(t *testing.T) {
+	s := createDdlSuite(t)
+	defer s.teardown(t)
+
 	// first add many data
 	workerNum := 10
 	base := *dataNum / workerNum
@@ -128,7 +133,7 @@ func (s *TestDDLSuite) TestColumn(c *C) {
 			defer wg.Done()
 			for j := 0; j < base; j++ {
 				k := base*i + j
-				s.execInsert(c, fmt.Sprintf("insert into test_column values (%d, %d)", k, k))
+				s.execInsert(fmt.Sprintf("insert into test_column values (%d, %d)", k, k))
 			}
 		}(i)
 	}
@@ -148,41 +153,42 @@ func (s *TestDDLSuite) TestColumn(c *C) {
 	updateDefault := int64(-2)
 	var alterColumn *table.Column
 
-	for _, t := range tbl {
-		c.Logf("run DDL %s", t.Query)
-		done := s.runDDL(t.Query)
+	for _, col := range tbl {
+		t.Run(col.Query, func(t *testing.T) {
+			done := s.runDDL(col.Query)
 
-		ticker := time.NewTicker(time.Duration(*lease) * time.Second / 2)
-		defer ticker.Stop()
-	LOOP:
-		for {
-			select {
-			case err := <-done:
-				c.Assert(err, IsNil)
-				break LOOP
-			case <-ticker.C:
-				count := 10
-				s.execColumnOperations(c, workerNum, count, &rowID, updateDefault)
+			ticker := time.NewTicker(time.Duration(*lease) * time.Second / 2)
+			defer ticker.Stop()
+		LOOP:
+			for {
+				select {
+				case err := <-done:
+					require.NoError(t, err)
+					break LOOP
+				case <-ticker.C:
+					count := 10
+					s.execColumnOperations(t, workerNum, count, &rowID, updateDefault)
+				}
 			}
-		}
 
-		if t.Add {
-			s.checkAddColumn(c, rowID, t.Default, updateDefault)
-		} else {
-			s.checkDropColumn(c, rowID, alterColumn, updateDefault)
-		}
+			if col.Add {
+				s.checkAddColumn(t, rowID, col.Default, updateDefault)
+			} else {
+				s.checkDropColumn(t, rowID, alterColumn, updateDefault)
+			}
 
-		tbl := s.getTable(c, "test_column")
-		alterColumn = table.FindCol(tbl.Cols(), t.ColumnName)
-		if t.Add {
-			c.Assert(alterColumn, NotNil)
-		} else {
-			c.Assert(alterColumn, IsNil)
-		}
+			tbl := s.getTable(t, "test_column")
+			alterColumn = table.FindCol(tbl.Cols(), col.ColumnName)
+			if col.Add {
+				require.NotNil(t, alterColumn)
+			} else {
+				require.Nil(t, alterColumn)
+			}
+		})
 	}
 }
 
-func (s *TestDDLSuite) execColumnOperations(c *C, workerNum, count int, rowID *int64, updateDefault int64) {
+func (s *ddlSuite) execColumnOperations(t *testing.T, workerNum, count int, rowID *int64, updateDefault int64) {
 	var wg sync.WaitGroup
 	// workerNum = 10
 	wg.Add(workerNum)
@@ -191,47 +197,58 @@ func (s *TestDDLSuite) execColumnOperations(c *C, workerNum, count int, rowID *i
 			defer wg.Done()
 			for j := 0; j < count; j++ {
 				key := int(atomic.AddInt64(rowID, 2))
-				s.execInsert(c, fmt.Sprintf("insert into test_column (c1, c2) values (%d, %d)",
+				s.execInsert(fmt.Sprintf("insert into test_column (c1, c2) values (%d, %d)",
 					key-1, key-1))
-				s.exec(fmt.Sprintf("insert into test_column values (%d, %d, %d)", key, key, key))
-				s.mustExec(c, fmt.Sprintf("update test_column set c2 = %d where c1 = %d",
+				_, err := s.exec(fmt.Sprintf("insert into test_column values (%d, %d, %d)", key, key, key))
+				require.NoError(t, err)
+				s.mustExec(fmt.Sprintf("update test_column set c2 = %d where c1 = %d",
 					updateDefault, randomNum(key)))
-				s.exec(fmt.Sprintf("update test_column set c2 = %d, c3 = %d where c1 = %d",
+				_, err = s.exec(fmt.Sprintf("update test_column set c2 = %d, c3 = %d where c1 = %d",
 					updateDefault, updateDefault, randomNum(key)))
-				s.mustExec(c, fmt.Sprintf("delete from test_column where c1 = %d", randomNum(key)))
+				require.NoError(t, err)
+				s.mustExec(fmt.Sprintf("delete from test_column where c1 = %d", randomNum(key)))
 			}
 		}()
 	}
 	wg.Wait()
 }
 
-func (s *TestDDLSuite) TestCommitWhenSchemaChanged(c *C) {
-	s.mustExec(c, "drop table if exists test_commit")
-	s.mustExec(c, "create table test_commit (a int, b int)")
-	s.mustExec(c, "insert into test_commit values (1, 1)")
-	s.mustExec(c, "insert into test_commit values (2, 2)")
+func TestCommitWhenSchemaChanged(t *testing.T) {
+	s := createDdlSuite(t)
+	defer s.teardown(t)
+
+	s.mustExec("drop table if exists test_commit")
+	s.mustExec("create table test_commit (a int, b int)")
+	s.mustExec("insert into test_commit values (1, 1)")
+	s.mustExec("insert into test_commit values (2, 2)")
 
 	s1, err := session.CreateSession(s.store)
-	c.Assert(err, IsNil)
+	require.NoError(t, err)
 	ctx := goctx.Background()
 	_, err = s1.Execute(ctx, "use test_ddl")
-	c.Assert(err, IsNil)
-	s1.Execute(ctx, "begin")
-	s1.Execute(ctx, "insert into test_commit values (3, 3)")
+	require.NoError(t, err)
+	_, err = s1.Execute(ctx, "begin")
+	require.NoError(t, err)
+	_, err = s1.Execute(ctx, "insert into test_commit values (3, 3)")
+	require.NoError(t, err)
 
-	s.mustExec(c, "alter table test_commit drop column b")
+	s.mustExec("alter table test_commit drop column b")
 
 	// When this transaction commit, it will find schema already changed.
-	s1.Execute(ctx, "insert into test_commit values (4, 4)")
+	_, err = s1.Execute(ctx, "insert into test_commit values (4, 4)")
+	require.NoError(t, err)
 	_, err = s1.Execute(ctx, "commit")
-	c.Assert(terror.ErrorEqual(err, plannercore.ErrWrongValueCountOnRow), IsTrue, Commentf("err %v", err))
+	require.Truef(t, terror.ErrorEqual(err, plannercore.ErrWrongValueCountOnRow), "err: %v", err)
 }
 
-func (s *TestDDLSuite) TestForIssue24621(c *C) {
-	s.mustExec(c, "use test")
-	s.mustExec(c, "drop table if exists t")
-	s.mustExec(c, "create table t(a char(250));")
-	s.mustExec(c, "insert into t values('0123456789abc');")
+func TestForIssue24621(t *testing.T) {
+	s := createDdlSuite(t)
+	defer s.teardown(t)
+
+	s.mustExec("use test")
+	s.mustExec("drop table if exists t")
+	s.mustExec("create table t(a char(250));")
+	s.mustExec("insert into t values('0123456789abc');")
 	_, err := s.exec("alter table t modify a char(12) null;")
-	c.Assert(err.Error(), Equals, "[types:1265]Data truncated for column 'a', value is '0123456789abc'")
+	require.EqualError(t, err, "[types:1265]Data truncated for column 'a', value is '0123456789abc'")
 }
